Improve support for long MP4 files. Twice as long files can now play, and for those that still are too long, exit with a codec failure rather than freeze Rockbox.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@11471 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libm4a/m4a.h
@@ -53,7 +53,7 @@ typedef struct
     } *time_to_sample;
     uint32_t num_time_to_samples;
 
-    uint32_t *sample_byte_size;
+    uint16_t *sample_byte_size;
     uint32_t num_sample_byte_sizes;
 
     uint32_t codecdata_len;
